added more authentication and checks to see if it's a manager or not for some endpoints
