+++ config/routes.rb
@@ -4,9 +4,9 @@ Rails.application.routes.draw do
   post 'sessions/logged_in'
   post 'reimbursement/create'
   get 'reimbursement/index'
-  get 'reimbursement/show'
+  get 'reimbursement/show/:id', to: 'reimbursement#show'
   get 'reimbursement/delete'
-  get 'reimbursement/update'
+  put 'reimbursement/update/:id', to: 'reimbursement#update'
   post 'user/create'
   get 'user/show'
   get 'user/index'
